🚚 move: restructure project files into organized directories
Reorganize source files into logical directories for better maintainability:
- core/ for game logic
- components/ for UI components
- styles/ for CSS
- types/ for TypeScript definitions
- assets/ for static files
