+++ queens-game/src/core/board.ts
