+++ queens-game/src/styles/style.css
