+++ queens-game/src/types/vite-env.d.ts
